perf: disable flush on snapshot saving
Firecracker is taking significant time on saving the snapshot file. For
comparison: writing the memory snapshot by us (1GiB) takes 1s, and
writing the snapshot file by Firecracker (150KiB) takes 8.5s. We presume
this is because of the syncing of the metadata, which is why we disable
it for now. The filesystem should still handle persisting, and given the
fact that the snapshot/memory files are temporary anyway, we can afford
to not sync/flush.